requirements.txt
google-generativeai
numpy
numexpr
flask
llama-index
python-dotenv